Don't stop processing TIFF tags when encountering an invalid tag type.
In some cases it _may_ be possible to continue reading tags. The entries
are fixed width, so it's possible to jump directly to the next one,
having stored an error.

However if we have 'left' the IFD (or perhaps were not in one all along)
then we might happily produce hundreds of garbage tags. This implementation
tracks the number of such errors and stops after some threshold.

Relates to #28.