Consolidate Exif description to new base class.
Removes duplicate functions for string descriptions of tag values, and
increases coverage of across all Exif directory types. Tags are not, in
practice, restricted to any particular directory type, and so this
approach gives much better coverage as proven by the regression tests
over the image database, as summarised in #82.

Closes #83.
Closes #8.
Closes #9.
Closes #3.