Introduce TiffReaderContext
We have been steadily adding parameters to members of ITiffHandler. We are about to add more to support GeoTIFF. Rather than do that, introduce a new type that encapsulates this common state. This makes the API simpler, and means we can add members in future without it being a breaking change.